+++ src/personal_cic/__init__.py
@@ -1,2 +1,2 @@
 """Personal CIC."""
-__version__ = "0.3.4"
+__version__ = "0.3.5"